feat: extract build auth redirect url to app gen
