+++ apps/api/src/app/auth/auth.controller.ts
@@ -37,6 +37,7 @@ import { ApiExcludeController, ApiTags } from '@nestjs/swagger';
 import { PasswordResetBodyDto } from './dtos/password-reset.dto';
 import {
   AuthService,
+  buildOauthRedirectUrl,
   SwitchEnvironment,
   SwitchEnvironmentCommand,
   SwitchOrganization,
@@ -82,58 +83,9 @@ export class AuthController {
   @Get('/github/callback')
   @UseGuards(AuthGuard('github'))
   async githubCallback(@Req() request, @Res() response) {
-    if (!request.user || !request.user.token) {
-      return response.redirect(`${process.env.FRONT_BASE_URL + '/auth/login'}?error=AuthenticationError`);
-    }
-
-    const url = this.buildOauthRedirectUrl(request);
+    const url = buildOauthRedirectUrl(request);
 
     return response.redirect(url);
-  }
-
-  private buildOauthRedirectUrl(request): string {
-    let url = process.env.FRONT_BASE_URL + '/auth/login';
-    const redirectUrl = JSON.parse(request.query.state).redirectUrl;
-
-    /**
-     * Make sure we only allow localhost redirects for CLI use and our own success route
-     * https://github.com/novuhq/novu/security/code-scanning/3
-     */
-    if (redirectUrl && redirectUrl.startsWith('http://localhost:') && !redirectUrl.includes('@')) {
-      url = redirectUrl;
-    }
-
-    url += `?token=${request.user.token}`;
-
-    if (request.user.newUser) {
-      url += '&newUser=true';
-    }
-
-    /**
-     * partnerCode, next and configurationId are required during external partners integration
-     * such as vercel integration etc
-     */
-    const partnerCode = JSON.parse(request.query.state).partnerCode;
-    if (partnerCode) {
-      url += `&code=${partnerCode}`;
-    }
-
-    const next = JSON.parse(request.query.state).next;
-    if (next) {
-      url += `&next=${next}`;
-    }
-
-    const configurationId = JSON.parse(request.query.state).configurationId;
-    if (configurationId) {
-      url += `&configurationId=${configurationId}`;
-    }
-
-    const invitationToken = JSON.parse(request.query.state).invitationToken;
-    if (invitationToken) {
-      url += `&invitationToken=${invitationToken}`;
-    }
-
-    return url;
   }
 
   @Get('/refresh')
