+++ packages/application-generic/src/services/auth/index.ts
@@ -1,0 +1,2 @@
+export * from './auth.service';
+export * from './shared';
